- Fixed SQL injection vulnerability with add/remove friend.

+++ app/Http/Controllers/ChatController.php
@@ -103,12 +103,17 @@ class ChatController extends Search\SearchController
 
     public function remove_add_Friend($friend_id)
     {
-        if(helper_isFriend($friend_id))
-            DB::table('friends')->where(['User_ID' => Auth::user()->id, 'Friend_ID' => $friend_id])->delete();
-        else
-            DB::table('friends')->insert(['User_ID' => Auth::user()->id, 'Friend_ID' => $friend_id]);
+        if (helper_UserExist($friend_id))
+        {
+            if(helper_isFriend($friend_id))
+                DB::table('friends')->where(['User_ID' => Auth::user()->id, 'Friend_ID' => $friend_id])->delete();
+            else
+                DB::table('friends')->insert(['User_ID' => Auth::user()->id, 'Friend_ID' => $friend_id]);
 
-        return redirect("/profile/".helper_GetUsernameById($friend_id));
+            return redirect("/profile/".helper_GetUsernameById($friend_id));
+        }
+
+        return redirect()->back();
     }
 
     //Used to add entries to the database showing the user has subscribed
